Refactor SLC to reuse CCA infrastructure
