+++ src/client/Microsoft.Identity.Client/ApiConfig/Executors/ManagedIdentityExecutor.cs
@@ -8,7 +8,7 @@ using Microsoft.Identity.Client.ApiConfig.Parameters;
 using Microsoft.Identity.Client.Instance.Discovery;
 using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.Internal.Requests;
-using Microsoft.Identity.Client.ManagedIdentity;
+using Microsoft.Identity.Client.ManagedIdentity.SLC;
 using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.ApiConfig.Executors
@@ -40,10 +40,20 @@ namespace Microsoft.Identity.Client.ApiConfig.Executors
                 requestContext,
                 _managedIdentityApplication.AppTokenCacheInternal).ConfigureAwait(false);
 
-            var handler = new ManagedIdentityAuthRequest(
-                ServiceBundle,
-                requestParams,
+            // MSI factory logic - decide if we need to use the legacy or the new MSI flow
+
+            RequestBase handler = null;
+#if TRA
+            // May or may not be initialized, depending on the state of the machine
+            handler = CredentialBasedMsiAuthRequest.TryCreate(
+                ServiceBundle, 
+                requestParams, 
                 managedIdentityParameters);
+#endif
+            handler ??= new LegacyMsiAuthRequest(
+                    ServiceBundle,
+                    requestParams,
+                    managedIdentityParameters);
 
             return await handler.RunAsync(cancellationToken).ConfigureAwait(false);
         }
