+++ src/client/Microsoft.Identity.Client/Internal/Requests/LegacyMsiAuthRequest.cs
@@ -1,25 +1,37 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 using System.Net;
+using System.Net.Http;
+using System.Security.AccessControl;
+using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.ApiConfig.Parameters;
+using Microsoft.Identity.Client.AppConfig;
 using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.Credential;
+using Microsoft.Identity.Client.Http;
 using Microsoft.Identity.Client.ManagedIdentity;
 using Microsoft.Identity.Client.OAuth2;
+using Microsoft.Identity.Client.Platforms.Features.KeyMaterial;
 using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.Internal.Requests
 {
-    internal class ManagedIdentityAuthRequest : RequestBase
+    /// <summary>
+    /// Old implementation of MSI, which works by getting tokens directly from the MSI infrastructure. 
+    /// </summary>
+    internal class LegacyMsiAuthRequest
+        : RequestBase
     {
         private readonly AcquireTokenForManagedIdentityParameters _managedIdentityParameters;
         private static readonly SemaphoreSlim s_semaphoreSlim = new SemaphoreSlim(1, 1);
 
-        public ManagedIdentityAuthRequest(
+        public LegacyMsiAuthRequest(
             IServiceBundle serviceBundle,
             AuthenticationRequestParameters authenticationRequestParameters,
             AcquireTokenForManagedIdentityParameters managedIdentityParameters)
@@ -54,7 +66,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
                 logger.Info("[ManagedIdentityAuthRequest] Access token retrieved from cache for managed identity.");
 
                 try
-                {  
+                {
                     var proactivelyRefresh = SilentRequestHelper.NeedsRefresh(cachedAccessTokenItem);
 
                     // may fire a request to get a new token in the background when AT needs to be refreshed
@@ -91,7 +103,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
         }
 
         private async Task<AuthenticationResult> GetAccessTokenAsync(
-            CancellationToken cancellationToken, 
+            CancellationToken cancellationToken,
             ILoggerAdapter logger)
         {
             AuthenticationResult authResult;
@@ -111,7 +123,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
                 // Bypass cache and send request to token endpoint, when
                 // 1. Force refresh is requested, or
                 // 2. Proactively Refreshed
-                if (_managedIdentityParameters.ForceRefresh || 
+                if (_managedIdentityParameters.ForceRefresh ||
                     AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo == CacheRefreshReason.ProactivelyRefreshed)
                 {
                     logger.Info("[ManagedIdentityAuthRequest] Sending Token response to managed identity endpoint ...");
@@ -153,7 +165,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
         {
             await ResolveAuthorityAsync().ConfigureAwait(false);
 
-            ManagedIdentityClient managedIdentityClient = 
+            ManagedIdentityClient managedIdentityClient =
                 new ManagedIdentityClient(AuthenticationRequestParameters.RequestContext);
 
             ManagedIdentityResponse managedIdentityResponse =
