+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/AbstractManagedIdentity.cs
