+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/AppServiceManagedIdentitySource.cs
