+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/AzureArcManagedIdentitySource.cs
