+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/CloudShellManagedIdentitySource.cs
