+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/EnvironmentVariables.cs
