+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ImdsManagedIdentitySource.cs
