+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ManagedIdentityErrorResponse.cs
