+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ManagedIdentityRequest.cs
