+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ManagedIdentityResponse.cs
