+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ManagedIdentitySource.cs
