+++ src/client/Microsoft.Identity.Client/LegacyManagedIdentity/ServiceFabricManagedIdentitySource.cs
