+++ src/client/Microsoft.Identity.Client/Microsoft.Identity.Client.csproj
@@ -117,8 +117,8 @@
     <Compile Include="$(PathToMsalSources)\**\*.cs" Exclude="$(PathToMsalSources)\obj\**\*.*" />
     <Compile Remove="$(PathToMsalSources)\Platforms\**\*.*;$(PathToMsalSources)\Resources\*.cs" />
     <Compile Remove="$(PathToMsalSources)\PlatformsCommon\PlatformNotSupported\ApiConfig\SystemWebViewOptions.cs" />
+   
     <Compile Remove="$(PathToMsalSources)\Credential\**\*.cs" />
-    <Compile Remove="$(PathToMsalSources)\ManagedIdentity\CredentialSource.cs" />
     <Compile Remove="$(PathToMsalSources)\Platforms\Features\KeyMaterial\**\*.cs" />
     <EmbeddedResource Include="$(PathToMsalSources)\Properties\Microsoft.Identity.Client.rd.xml" />
   </ItemGroup>
